+++ src/themes/two/src/components/BusinessProductAddForm/styles.js
@@ -1,0 +1,108 @@
+import styled, { css } from 'styled-components'
+
+export const FormInput = styled.form`
+  margin-top: 20px;
+  overflow: auto;
+  width: 100%;
+  padding: 20px;
+  box-sizing: border-box;
+`
+
+export const ProductImage = styled.div`
+  width: 200px;
+  height: 200px;
+  border-radius: 8px;
+  overflow: hidden;
+  cursor: -webkit-grab;
+  cursor: grab;
+  filter: drop-shadow(0px 1px 4px rgba(0, 0, 0, 0.1));
+
+  img,
+  div {
+    width: 100%;
+    border-radius: 8px;
+    height: 100%;
+    overflow: hidden;
+  };
+
+  img{
+    object-fit: cover;
+  }
+`
+
+export const SkeletonWrapper = styled.div`
+  span{
+    height: 100%;
+    position: relative;
+    top: -2.5px;
+  }
+`
+export const UploadImageIconContainer = styled.div`
+  position: absolute;
+  top: 0px;
+  background: rgba(0,0,0,0.2);
+  ${({ small }) => small ? css`
+    padding: 4px;
+  ` : css`
+    padding: 8px;
+  `}
+`
+
+export const UploadImageIcon = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  color: #FFF;
+  border: 1px dashed #fff;
+  span {
+    margin: 0;
+    text-align: center;
+    font-size: 14px;
+  }
+
+  svg {   
+    ${({ small }) => small ? css`
+      width: 25px;
+      height: 25px;
+    ` : css`
+      width: 45px;
+      height: 45px;
+    `}
+  }
+`
+
+export const InputWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  margin-top: 24px;
+
+  label {
+    color: ${props => props.theme.colors.headingColor};
+    margin-bottom: 10px;
+  }
+
+  input,
+  textarea {
+    color: ${props => props.theme.colors.headingColor};
+  }
+
+  input::-webkit-outer-spin-button,
+  input::-webkit-inner-spin-button {
+    -webkit-appearance: none;
+    margin: 0;
+  }
+
+  input[type=number] {
+    -moz-appearance: textfield;
+  }
+`
+
+export const ActionsForm = styled.div`
+  margin: 30px 0;
+  position: sticky;
+  top: 100%;
+  > button {
+    height: 42px;
+  }
+`
