+++ src/themes/two/src/components/SingleBusinessProduct/styles.js
@@ -160,6 +160,10 @@ export const ActionSelectorWrapper = styled.div`
       }
     }
   }
+
+  .dropdown-item {
+    font-size: 14px;
+  }
 `
 
 export const ProductTypeImage = styled.div`
